--- src/main/java/com/back/together02be/stock/cache/StockPriceCache.java
@@ -1,7 +1,0 @@
-package com.back.together02be.stock.cache;
-
-public record StockPriceCache(
-        Long currentPrice,
-        Double changeRate
-) {
-}
